requirements.txt
# Install with: 'pip install -r requirements.txt'

# Infrastructure

Django == 1.7
djangorestframework==3.0.1
redis==2.10.3
psycopg2==2.5.4
webstack-django-sorting==0.4.1
celery==3.1.15
wsgiref==0.1.2
linaro-django-pagination==2.0.2
gunicorn==19.1.1
